+++ source/parameters/base/base.h
@@ -29,8 +29,8 @@ class Base
  double _minStdDevChange;
 
  // Constructor / Destructor
- Base(nlohmann::json& js, int seed);
- Base(int seed);
+ Base(nlohmann::json& js, size_t seed);
+ Base(size_t seed);
  ~Base();
 
  virtual double getDensity(double x) = 0;
